Move pytorch hooks to a standalone file
